Update ban command to check for open tickets
The number of message-delete days for the ban command has been updated to be dependent on whether the user has any open tickets. If the user has any open tickets, the number of delete days is set to 0. Otherwise, it is set to 7. This has been applied across all instances of ban commands in moderation.