[Botapplication] Update method to async and add WebUI status check
The code has been adjusted to use the asynchronous version of the RunAsp method. Additionally, a new public static class has been added to hold temporary variables, starting with a boolean flag to check whether the WebUI application is running or not. This introduces an extra layer of control over the WebUI status.