Fix method used to retrieve user's display name
The commit changes the method used to retrieve the user's display name in the UserService from RetrieveRole to RetrieveDisplayName for better representation of data. It also adds a new method in the AuthUtils helper to fetch the display name based on user id. These adjustments enable more accurate retrieval and representation of the user's display name.