Add method to get UserId and modify roles in authorization
A method is added to the UserService that retrieves the UserId from the HttpContext. This is utilized in the debug page and the NavMenu shared page for better user identification. Furthermore, the roles allowed in the debug page were updated and the initialization logic in the navigation menu was modified to only show up once the user is identified.