Add SkiaSharp package and improve user rank visuals
This commit adds the SkiaSharp package to the project for image creation. It also significantly enhances the display of rank and experience point progression in the Rank command. This is achieved by creating a visual progress bar, detailing overall XP and user rank. Finally, the LevelUtils class has been expanded with various functions to better calculate user level and XP data.