Add user level recalculation feature
A new function 'RecalculateUserLevel' has been added to the LevelUtils class which recalculates a user's level based on their experience points and updates it in the database. The 'RankCommand' class has been updated to call this function whenever a user's rank is retrieved. Additionally, a function to recalculate all users' levels has been included.