Rearrange role and level display in ShowConfigCommand
The DisplayOrder in ShowConfigCommand.cs has been updated to prioritize leveling followed by role naming. This change presents this information in a clearer and more coherent manner, first showing the level and then the corresponding role. An extra line is also added to differentiate between different categories of information in the output.