Add experience modification methods and commands
Three new methods have been added to the LevelUtils module for adjusting a DiscordUser's experience: SetXp, AddXp, and RemoveXp. These are evenly complemented with a new interaction command adjusting the experience of a user. Additionally, a ModifyAction enumeration was created to specify the type of XP adjustment action in the new interaction command.