Add channel blacklist management commands
A new command file ManageBlockedChannelsCommand.cs has been added to enable the management of channel blacklists. It includes commands to add and remove Discord channels to and from a blacklist. Additional utility methods have been also added in LevelUtils.cs to check if a channel is blacklisted and to add or remove a channel from the blacklist in the database.