Add level multiplier management functionality
A new feature was implemented to manage the level multiplier in the level system settings. This function allows the multiplier for a given level type to be set and set as active or inactive, as well as responding to the user with the successful operation's output. This will enhance the flexibility and customization of the leveling system within the application.