Replace TicketDatabaseService with DatabaseService
The TicketDatabaseService file has been deleted and all its references have been replaced with DatabaseService. This refactoring change was made throughout various files and was aimed at streamlining the access to the database by means of a consolidated service provider, NpgsqlDataSource. This should improve code maintainability and consistency in the project.