Add change level card button to rank response
The "RankCommand.cs" file was modified to include a new button in the response to a rank command. This button, labeled "Hintergrund ändern (bald verfügbar)" with an emoji, will take the user to a website where they can change their level card. The button is currently disabled, as the functionality is not yet available.