Refactor leaderboard data retrieval and caching
This update moves the leaderboard data handling from the Leaderboard component into LevelUtils. A new list for cached leaderboard data has been added. The process now includes checking user presences on the server and in cache, and falling back on a default user view if they are not found. This change removes redundant code in the leaderboard component and centralizes the leaderboard data management.