Change List to HashSet in LoadLeaderboardData
In LevelUtils.cs, the type of tempLeaderboardData used to load leaderbard data has been changed from List to HashSet, to enhance performance. Additionally, the logic of clearing the tempLeaderboardData has been altered: now, the data from HashSet is converted to a list and is assigned to cachedLeaderboardData without clearing the temporary leaderboard data.