Add build date and time tracking features
This update adds new functionalities to the AGC Management project by incorporating the ability to track the build date and time of the assembly. It provides two methods to obtain this information, either as a DateTime or as a Unix timestamp. The commit also incorporates these timestamps into the StatsCommand displaying the uptime and compile date.