Refactor CachingService to use database instead of files
The CachingService has been refactored to use a database instead of files, with associated changes made to related services and commands. This includes the renaming of Enum FileCacheType to CustomDatabaseCacheType, adjusting parameters and creating of new database table for caching. This change streamlines data access and potentially improves cache performance.