Update UI and functionality in AdjustRankCard
Css styles for input elements and buttons have been revised for better usability in Pages/UserSettings/LevelSystem/AdjustRankCard.razor.css. For AdjustRankCard.razor, font selection is now done through a dropdown, and user feedback is added for file upload. Upload file size is now also verified and user is informed if the file size is too large or an incompatible format is chosen.