Add user ID retrieval and web login logging functionality
The update introduces a method to retrieve user IDs from JSON property in AuthUtils. Also, a logging utility for Web OAuth Discord logins is implemented in LoggingUtils. This development is critical for recording log in activities, thereby enhancing security. A new table 'dashboardlogins' is created in the DatabaseService to save the logs. A new LoginLogs.razor page is also introduced to visualize these logs. Access to the logs and errors are provided through Bot Logs links in the Logging Menu.