Refactor layout in TeamArea and LoggingMenu razor pages
In this commit, the layout structuring of TeamArea.razor and LoggingMenu.razor pages has been improved. The tag content in TeamArea.razor has been reordered for better readability. In LoggingMenu.razor, the positioning of cards has been reworked which also led to a clearer overall structure and easier to maintain code.