Remove 'isbot' field from metrics tables
The 'isbot' field has been removed from 'metrics_messages' and 'metrics_voice' tables in the database. This change was also reflected in our 'DatabaseService.cs' and various event handlers to prevent them from inserting this non-existent field into the database. Now, the code directly checks if the event is triggered by a bot and, if so, it exits the function immediately.