Remove application submission process from OnApplyComponentInteraction
The application submission functionality that was previously running as a Task within OnApplyComponentInteraction method has been removed. This included the creation and insertion of an application entry into a Postgres database. This decision was motivated by a restructuring of the codebase, aiming to delegate this operation elsewhere for code maintainability and organization.