Update DisCatSharp packages and modify return values and properties
This commit updates the version of DisCatSharp packages in the AGC Management project. It also changes the return value from a hardcoded integer to a method call converting days to seconds in Helpers.cs. Additionally, a change has been made to use a property instead of a method to get the bot owner information in the StatsCommand.cs class.